+++ logger.py
@@ -1,8 +1,5 @@
-import functools
 import logging
 import socket
-import warnings
-
 
 def get_logger(name):
     logger = logging.getLogger(name)
@@ -23,25 +20,6 @@ class ContextFilter(logging.Filter):
         return True
 
 #warnings.simplefilter('always', DeprecationWarning)
-
-class deprecated(object):
-    def __init__(self, msg=None):
-        """
-        A decorator which can be used to mark functions
-        as deprecated.It will result in a deprecation warning being shown
-        when the function is used.
-        """
-        self.msg = msg
-
-    def __call__(self, func):
-        message = self.msg or "Use of deprecated function '{}`.".format(func.__name__)
-
-        @functools.wraps(func)
-        def wrapper_func(*args, **kwargs):
-            warnings.warn(message, DeprecationWarning, stacklevel=2)
-            return func(*args, **kwargs)
-        return wrapper_func
-
 
 # build context filter
 _context_filter = ContextFilter(socket.gethostname())
